requirements.txt
aws-cdk-lib==2.86.0
constructs>=10.0.0,<11.0.0
pyyaml==6.0
cdk-nag==2.27.60